GPEC/requirements.txt
absl-py==1.4.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
anyio==3.7.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
attrs==21.4.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
chart-studio==1.1.0
chembl-webresource-client==0.10.8
click==8.1.3
cloudpickle==3.0.0
cmake==3.27.7
colorama==0.4.6
colorlover==0.3.0
comm==0.1.3
comtypes==1.1.14
contourpy==1.1.0
cufflinks==0.17.3
cycler==0.11.0
Cython==3.0.9
DateTime==5.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
dlib @ file:///C:/Users/keane/Downloads/dlib-19.22.99-cp310-cp310-win_amd64.whl#sha256=f2181b6724669fb0147c9ebc326d72b94de41bced1812eff5df4b133b5d0b575
easydict==1.12
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastjsonschema==2.17.1
featuretools==1.30.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.40.0
fqdn==1.5.1
frozendict==2.3.8
frozenlist==1.4.1
fsspec==2023.10.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.56.2
gTTS==2.3.1
h11==0.14.0
h5py==3.9.0
holidays==0.44
html5lib==1.1
idna==3.4
importlib-metadata==6.11.0
importlib_resources==6.1.3
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.context==4.3.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.13.1
keras-efficientnets==0.1.7
keyboard==0.13.5
kiwisolver==1.4.4
libclang==16.0.6
lightgbm==4.3.0
lstm==0.1.0
lxml==4.9.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
more-itertools==9.1.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
mysql-client==0.0.1
mysql-connector-python==8.0.32
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.2.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
np-utils==0.6.0
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.3.1
packaging==23.0
pandas==2.0.2
pandas-datareader==0.10.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pickleshare==0.7.5
Pillow==9.5.0
pip-autoremove==0.10.0
platformdirs==4.2.0
playwright==1.41.2
plotly==5.15.0
pmdarima==2.0.4
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.53
pycparser==2.21
pydeck==0.8.1b0
pydot==1.4.2
pyee==11.0.1
PyGetWindow==0.0.9
Pygments==2.15.1
PyMsgBox==1.0.9
PyMySQL==1.0.3
pyparsing==3.1.0
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.7
pyttsx3==2.90
pytweening==1.0.4
pytz==2023.3
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
rdkit-pypi==2022.9.5
regex==2023.6.3
requests==2.28.2
requests-cache==0.7.5
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
selenium==4.7.2
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
statsmodels==0.14.1
streamlit==1.29.0
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tk==0.1.0
tkcalendar==1.6.1
toml==0.10.2
toolz==0.12.0
torch==2.1.1
torch_geometric==2.5.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.3.2
tqdm==4.64.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
url-normalize==1.4.3
urllib3==1.26.14
validators==0.22.0
virtualenv==20.25.1
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==1.6.0
Werkzeug==2.3.6
widgetsnbextension==4.0.7
wolframalpha==5.0.0
woodwork==0.29.0
wrapt==1.15.0
wsproto==1.2.0
xgboost==2.0.2
xmltodict==0.13.0
yarl==1.9.4
yfinance==0.2.22
zipp==3.17.0
zope.interface==6.0
